Разбила main.js на модули
