+++ src/view/load-more-button.js
@@ -1,0 +1,14 @@
+import {boardElement} from "./board";
+import {render} from "../main";
+
+const createLoadMoreButtonTemplate = () => {
+  return (
+    `
+<button class="load-more" type="button">load more</button>
+    `
+  );
+};
+
+render(boardElement, createLoadMoreButtonTemplate(), `beforeend`);
+
+export {createLoadMoreButtonTemplate};
